Fix yay, polybar and xinitrc

+++ install.sh
@@ -4,13 +4,13 @@
 
 name="dabljues"
 
-necessary_binaries=( yay python-pywal i3-gaps git zsh pulseaudio )
+necessary_binaries=( python-pywal i3-gaps git zsh pulseaudio )
 pacman_binaries_file="lists/progs.txt"
 aur_binaries_file="lists/aur_progs.txt"
 
 # Copying env file
 cp -f dotfiles/.config/env ~/.config/env
-mv ~/.dir_colors ~/.config/dircolors
+[ -d ~/.dir_colors ] && mv ~/.dir_colors ~/.config/dircolors
 source ~/.config/env
 
 mkdir -p $XDG_DATA_HOME
@@ -26,6 +26,16 @@ echo "Updating the system!"
 
 #sudo pacman-mirrors -g >/dev/null 2>&1
 #sudo pacman -Syyu --noconfirm >/dev/null 2>&1
+
+# Installing git
+
+sudo pacman --noconfirm --needed -S git >/dev/null 2>&1
+
+# Installing yay
+
+echo "Installing yay"
+
+[ -z "$(command -v yay)" ] && git clone https://aur.archlinux.org/yay.git && cd yay && sudo -u "$name" makepkg --noconfirm -si >/dev/null 2>&1 && cd .. && rm -rf yay
 
 # Necessary binaries
 echo "Installing necessary binaries"
@@ -55,7 +65,7 @@ done < $aur_binaries_file
 wal -i images/ocean.jpg
 
 # oh-my-zsh
-rm ~/.zshrc
+[ -f ~/.zshrc ] && rm ~/.zshrc
 mkdir -p $XDG_CONFIG_HOME/zsh
 curl -Lo install_zsh.sh https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh
 # replace paths in oh-my-zsh installer
@@ -70,13 +80,13 @@ rm install_zsh.sh
 echo "Copying dotfiles"
 
 rm -rf ~/.dmenu*
-rm ~/.profile
+[ -f ~/.profile ] && rm ~/.profile
 rm ~/.bash*
-rm -rf ~/.moc
-rm ~/.zshrc
-rm ~/.Xresources
+[ -d ~/.moc ] && rm -rf ~/.moc
+[ -f ~/.zshrc ] && rm ~/.zshrc
+[ -f ~/.Xresources ] && rm ~/.Xresources
 rm ~/.xsession*
-rm -rf ~/.i3
+[ -d ~/.i3 ] && rm -rf ~/.i3
 yes | cp -rfT dotfiles /home/$USER >/dev/null 2>&1
 
 # Most important command! Get rid of the beep!
@@ -88,23 +98,23 @@ sudo chsh -s $( which zsh ) $USER >/dev/null 2>&1
 sudo -u "$name" mkdir -p "/home/$name/.cache/zsh/"
 
 # Home clean-up
-mv ~/.gnupg $XDG_DATA_HOME
+[ -f ~/.zshrc ] && mv ~/.gnupg $XDG_DATA_HOME
 
 mkdir -p ~/.config/gtk-2.0
-mv ~/.gtkrc-2.0 $GTK2_RC_FILES
+[ -d ~/.gtkrc-2.0 ] && mv ~/.gtkrc-2.0 $GTK2_RC_FILES
 
 mkdir -p ~/.config/vim
 mv ~/.vim* ~/.config/vim
 
-mv ~/.Xauthority ~/.config/Xauthority
+[ -f ~/.Xauthority ] && mv ~/.Xauthority ~/.config/Xauthority
 #mv ~/.mozilla ~/.config/mozilla
 
-mv ~/.thumbnails ~/.config/thumbnails
-mv ~/.urxvt ~/.config/urxvt
+[ -d ~/.thumbnails ] && mv ~/.thumbnails ~/.config/thumbnails
+[ -d ~/.urxvt ] && mv ~/.urxvt ~/.config/urxvt
 
-mv ~/.gitconfig ~/.config/git/config
+[ -d ~/.gitconfig ] && mv ~/.gitconfig ~/.config/git/config
 
-mv ~/.Xclients ~/.config/Xclients
+[ -d ~/.Xclients ] && mv ~/.Xclients ~/.config/Xclients
 
 # Restart pulseaudio
 killall pulseaudio
